TypeError: Cannot read property voteScore of undefined
